% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitplot.R
\name{pitplot}
\alias{pitplot}
\title{Plot pit stop results}
\format{
Tibble
}
\usage{
pitplot(pits_data, type = 3, title_text = NULL)
}
\arguments{
\item{pits_data}{Tibble data generated by the pits() function}

\item{type}{Plot type: individual pit stop by driver (1), grouped by team (2), grouped by driver (3, by default)}

\item{title_text}{Text for the plot title, in quotes (" ") (if is omitted, a default text will be used).}
}
\value{
A ggplot object
}
\description{
Plot pit stop results (MUST BE in tibble format)
}
