% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pits.R
\name{pits}
\alias{pits}
\title{Pit stops of a race or set}
\usage{
pits(round, year)
}
\arguments{
\item{round}{Number of the race (integer), set of races a:b (vector) o "all" (character)}

\item{year}{Year of the race (integer). Available: ≥ 2025}
}
\value{
A tibble containing the pit stops values of the specified race(s)
}
\description{
Pit stop results of a race or set
}
\examples{
pits(10, 2025)
pits(1:2, 2025)
pits("all",2025)
}
